IN-43: CreateNewPasswordPage, CreateNewPasswordForm refactor
