+++ src/features/auth/reset-confirm/ui/CreateNewPasswordForm/CreateNewPasswordForm.module.scss
