--- src/pages/auth/CreateNewPasswordPage/CreateNewPasswordPage.tsx
@@ -1,10 +1,0 @@
-import { CreateNewPasswordForm } from '@/src/features/auth/reset-confirm/ui/CreateNewPasswordForm';
-import Card from '@/src/shared/ui/Card/Card';
-
-export default function CreateNewPasswordPage() {
-  return (
-    <Card>
-      <CreateNewPasswordForm />
-    </Card>
-  );
-}
